% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-Format2.R, R/format-rndr2.R
\name{Format2}
\alias{Format2}
\alias{ergaenze_vector}
\alias{drop_0_leading}
\alias{make_format}
\alias{Format2.matrix}
\alias{Format2.tbl_df}
\alias{Format2.data.frame}
\alias{Format2.list}
\alias{Format2.default}
\alias{rndr2}
\alias{rndr2.matrix}
\alias{rndr2.data.frame}
\alias{rndr2.default}
\title{Render numbers}
\usage{
Format2(x, ...)

ergaenze_vector(x, n_cols = NULL)

drop_0_leading(x, OutDec = getOption("OutDec"))

make_format(x, digits, format, drop0trailing, drop0leading, decimal.mark,
  na.strings = "NA", na.symbol = "")

\method{Format2}{matrix}(x, digits = 2, lead.zero = TRUE,
  drop0leading = !lead.zero, format = "f", ...)

\method{Format2}{tbl_df}(x, ...)

\method{Format2}{data.frame}(x, digits = 2, lead.zero = TRUE,
  drop0leading = !lead.zero, format = "f", ...)

\method{Format2}{list}(x, digits = NULL, lead.zero = TRUE,
  drop0leading = !lead.zero, format = "f", ...)

\method{Format2}{default}(x, digits = 2, lead.zero = TRUE,
  scientific = FALSE, nsmall = ifelse(is.null(digits), 0L, digits),
  drop0leading = !lead.zero, drop0trailing = FALSE, format = "f",
  decimal.mark = getOption("OutDec"), ...)

rndr2(...)

\method{rndr2}{matrix}(x, ...)

\method{rndr2}{data.frame}(x, ...)

\method{rndr2}{default}(x, digits = 2, format = "f", drop0leading = FALSE,
  drop0trailing = FALSE, decimal.mark = getOption("OutDec"),
  include.leading = FALSE, include.trailing = FALSE,
  include.bracket = FALSE, symbol.bracket = c("(", ")"),
  symbol.leading = " # ", symbol.trailing = " # ", ...)
}
\arguments{
\item{x}{Objekt kann Vector data.frame oder list sein}

\item{...}{an FormatC}

\item{n_cols}{anzahl an Elementen  (intern in ergaenze_vector())}

\item{OutDec}{Detzimal-Trenzeichen wird bei rndr_ verwendet und ueber Optionen gesteuert}

\item{digits}{Nachkommastellen als vektor oder liste}

\item{format}{format an formatC()}

\item{drop0trailing}{Nullen am Ende an formatC()}

\item{decimal.mark}{Kommatrennzeichen an formatC()}

\item{na.strings, na.symbol}{in der internen Funktion make_format genutzt und kann nur über Optionen geändert werden (noch nicht implementiert)}

\item{lead.zero, drop0leading}{Null am  Anfang an formatC()}

\item{scientific, nsmall}{format an formatC()}

\item{include.leading, symbol.leading}{Vorangestelltes Symbol}

\item{include.trailing, symbol.trailing}{Nachgestelltes Symbol}

\item{include.bracket, symbol.bracket}{Klammern}
}
\value{
Gleiches objekt wie der Input aber mit Character.
}
\description{
Formatieren von Zahlen nach APA-Style.

rndr2 ist die interne default-Methode um statistische 
Kennzahlen wie t, p, oder r zu Formatieren
}
\examples{

stp25rndr:::ergaenze_vector(c(1,2,4), 6)


stp25rndr:::drop_0_leading(0.26589)


stp25rndr:::make_format(c(12.5, NA, 32.1459),
                          2, "f", FALSE, FALSE, ".")



 x<- rnorm(10)
df <- data.frame(Item=c("a", "b", "c"),
                 x=x[1:3],
                 x2=c(1.2,2.3,.3),
                 beta=  c(.22,.13, NA),
                 x3=c(.42,.03,.003),
                 p.value=c(0.0002456,0.0398,.256))

mx1<- as.matrix(df[,-1])
mx2<-matrix(rnorm(10), ncol=2)
#- matrix ---------------------------
res <- Format2(mx1, digits=2)
cat("\\n in: matrix out:", class(res)," \\n")
res
x<-mx2[,1]
res <- Format2(x, digits=c(1:5), drop0trailing = TRUE)
cat("\\n in: ", class(x)," out:", class(res)," \\n")
res
Format2(x, c(1:3))


#- tbl_df/data.frame ----------------------
#Format2(tribble::tbl_df(df), digits=3)


 x<- rnorm(10)
df <- data.frame(Item=c("a", "b", "c"),
                 x=x[1:3],
                 x2=c(1.2,2.3,.3),
                 beta=  c(.22,.13, NA),
                 x3=c(.42,.03,.003),
                 p.value=c(0.0002456,0.0398,.256))
#- data.frame ----------------------
res <- Format2(df[,-1], digits=2, FALSE)
cat("\\n in: data.drame out:", class(res)," \\n")
res
Format2(df, digits=3)


lx<- list(a=1:5, b=rnorm(10))

#- list --------------------------
res <- Format2(lx, 2, FALSE)
cat("\\n in: list out:", class(res)," \\n")
res

 x<- rnorm(10)
 
 #- vector -------------------------
 res<-Format2(x, digits=2, FALSE)
 cat("\\n in: numeric out:", class(res)," \\n")
 res
 Format2(as.character(x), digits=3)
 Format2(factor(x), digits=3)

}
