% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format2.R, R/format-rndr2.R
\name{Format2}
\alias{Format2}
\alias{ergaenze_vector}
\alias{drop_0_leading}
\alias{make_format}
\alias{Format2.matrix}
\alias{Format2.tbl_df}
\alias{Format2.data.frame}
\alias{Format2.list}
\alias{Format2.default}
\alias{rndr2}
\alias{rndr2.matrix}
\alias{rndr2.data.frame}
\alias{rndr2.default}
\title{Render numbers}
\usage{
Format2(x, ...)

ergaenze_vector(x, n_cols = NULL)

drop_0_leading(x, OutDec = getOption("OutDec"))

make_format(x, digits, format, drop0trailing, drop0leading, decimal.mark,
  na.strings = "NA", na.symbol = "")

\method{Format2}{matrix}(x, digits = 2, lead.zero = TRUE, type = "digits",
  drop0leading = !lead.zero, format = if (type[1] == "digits") "f" else "g",
  ...)

\method{Format2}{tbl_df}(x, ...)

\method{Format2}{data.frame}(x, digits = 2, lead.zero = TRUE,
  type = "digits", drop0leading = !lead.zero, format = if (type[1] ==
  "digits") "f" else "g", ...)

\method{Format2}{list}(x, digits = NULL, lead.zero = TRUE,
  type = "digits", drop0leading = !lead.zero, format = if (type[1] ==
  "digits") "f" else "g", ...)

\method{Format2}{default}(x, digits = 2, lead.zero = TRUE,
  type = "digits", scientific = FALSE, nsmall = ifelse(is.null(digits),
  0L, digits), drop0leading = !lead.zero, drop0trailing = FALSE,
  format = if (type == "digits") "f" else "g",
  decimal.mark = getOption("OutDec"), ...)

rndr2(x, ...)

\method{rndr2}{matrix}(x, ...)

\method{rndr2}{data.frame}(x, ...)

\method{rndr2}{default}(x, digits = 2, format = "f", drop0leading = FALSE,
  drop0trailing = FALSE, decimal.mark = getOption("OutDec"),
  include.leading = FALSE, include.trailing = FALSE,
  include.bracket = FALSE, symbol.bracket = c("(", ")"),
  symbol.leading = " # ", symbol.trailing = " # ")
}
\arguments{
\item{x}{Objekt kann Vector data.frame oder list sein}

\item{...}{an FormatC}

\item{n_cols}{anzahl an Elementen  (intern in ergaenze_vector())}

\item{digits}{Nachkommastellen als vektor oder liste}

\item{drop0trailing}{Nullen am Ende an formatC()}

\item{decimal.mark}{Kommatrennzeichen an formatC()}

\item{na.strings, na.symbol}{in der internen Funktion make_format genutzt und kann nur über Optionen geändert werden (noch nicht implementiert)}

\item{lead.zero, drop0leading}{Null am  Anfang an formatC()}

\item{type, format}{format an formatC()}

\item{scientific, nsmall}{format an formatC()}

\item{include.leading, symbol.leading}{Vorangestelltes Symbol}

\item{include.trailing, symbol.trailing}{Nachgestelltes Symbol}

\item{include.bracket, symbol.bracket}{Klammern}
}
\value{
Gleiches objekt wie der Input aber mit Character.
}
\description{
Formatieren von Zahlen nach APA-Style.

rndr2 ist die interne default-Methode um statistisch e Kennzahlen wie t, p, oder r zu Formatieren
 rendr default
}
\examples{
#'
x<- rnorm(10)
df <- data.frame(Item=c("a", "b", "c"),
                 x=x[1:3],
                 x2=c(1.2,2.3,.3),
                 beta=  c(.22,.13, NA),
                 x3=c(.42,.03,.003),
                 p.value=c(0.0002456,0.0398,.256))

mx1<- as.matrix(df[,-1])
mx2<-matrix(rnorm(10), ncol=2)
lx<- list(a=1:5, b=rnorm(10))

# Tvl<- c(1.0145, 25.45)
# Df<- c(45,44.14)
# Pvl<-c(0.0045,0.547863)
# rndr_T(Tvl, Df, Pvl)
#
#
#
# digits = 2
# lead.zero = TRUE
# type = "digits"
# scientific = FALSE
# nsmall =  ifelse(is.null(digits), 0L,  digits)
#
# formatC(Tvl, digits = 2,
#         format = "fg",
#         decimal.mark = getOption("OutDec")
# )
#
#
# glue(' T<sub>({
#      formatC(Df, digits = 2,
#      format = "f",
#      decimal.mark = getOption("OutDec") )
#      })</sub>={
#      formatC(Tvl, digits = 2,
#      format = "f",
#      decimal.mark = getOption("OutDec"))
#      }, p={
#      formatC(Pvl, digits = 3,
#      format = "g", drop0trailing = TRUE,
#      decimal.mark = getOption("OutDec"))
#      }')
#
#
#

sprintf("\%.2f", x)
formatC(x, decimal.mark =",")
format(x, decimal.mark =",")
Format2(x, OutDec=",")
Format2(x, decimal.mark=",")
Format2(x )

Format2(df, digits=c(0,2,1,2,3,3),
        format=c("f","f","g","f","f","f"),
        drop0leading=c(F,F,F,F,F,T) )

Format2(c(0.345,2.45,5.1,2.67,39.332,.83), digits=c(0,2,1,2,3,3),
        format=c("f","f","g","f","f","f"),
        drop0leading=c(F,F,F,F,F,T) )



Format2(lx, list(c(0), c(1,1,4,3)))



stp25rndr:::ergaenze_vector(c(1,2,4), 6)


stp25rndr:::drop_0_leading(0.26589)


stp25rndr:::make_format(c(12.5, NA, 32.1459),
                          2, "f", FALSE, FALSE, ".")




#- matrix ---------------------------
res <- Format2(mx1, digits=2)
cat("\\n in: matrix out:", class(res)," \\n")
res
x<-mx2[,1]
res <- Format2(x, digits=c(1:5), drop0trailing = TRUE)
cat("\\n in: ", class(x)," out:", class(res)," \\n")
res
Format2(x, c(1:3))


#- tbl_df/data.frame ----------------------
Format2(tribble::tbl_df(df), digits=3)



#- data.frame ----------------------
res <- Format2(df[,-1], digits=2, FALSE)
cat("\\n in: data.drame out:", class(res)," \\n")
res
Format2(df, digits=3)




#- list --------------------------
res <- Format2(lx, 2, FALSE)
cat("\\n in: list out:", class(res)," \\n")
res



# #- vector -------------------------
# res<-Format2(x, digits=2, FALSE)
# cat("\\n in: numeric out:", class(res)," \\n")
# res
# Format2(as.character(x), digits=3)
# Format2(factor(x), digits=3)

}
